Add global mod setting 'proxyurl'.
Adds a new setting 'proxyurl' that can be used to set a proxy as hostname:port.
This will be used for communication with the Zoom API (but not anywhere else in Moodle).
This way, you can enable a proxy *only* for mod Zoom, and not for any other parts of Moodle.