Add admin setting to disable webinars - Fixes #188 (#200)
This patch adds an admin setting that allows the admin to control if the option to create a webinar is offered to users during the creation of a meeting or not.

This setting only affects the Moodle activity instance settings. Even if the admin decides to always show the option (which is the default and the Status Quo before this patch), the user will still need a valid license for webinars to host a webinar.

This setting is especially helpful to organizations who don't have webinar licenses or don't want to use them from within Moodle.