Add admin setting to suppress the "Download ical" button - Fixes #192
This patch adds an admin setting which simply lets the admin decide if he wants to hide or show the download iCal on each Zoom activity overview page.
Of course, if the download is disabled, the download of the iCal file is prevented even if such a user directly accesses the URL to download the file.