Make activity overview page more usable (#207)
* Cluster and improve activity overview page

This patch adds these changes to the activity overview page
* Move the "Meeting not found" notifications below the main page heading
* Split the large meeting information table and add headings between the table rows. This was done similar to what was done before on the activity instance settings page (with the only exception of the host information which was added to the "Schedule" section)
* Improve the appearance of the join button / unavailability notification

* Add admin setting to hide the all meetings link

This patch adds an admin setting which simply lets the admin decide if he wants to hide or show the all meetings link at the bottom of each Zoom activity overview page.

* Add admin setting to display the lead time

This patch adds an admin setting to let admins display the leadtime to the users. With this option, users are informed that / when they can join the meeting before the scheduled start tim
e. This might keep users from constantly reloading the page until they can join.