Add a meeting capacity warning - Fixes #187
This patch adds an admin setting with which the admin can show a warning notification if there are more active and enrolled participants in the course than the host's Zoom license meeting capacity is. The notification will be shown to the host (and alternative hosts) on the Zoom activity overview page. It will recommend the host to turn to the Zoom account owner to obtain a larger Zoom license if necessary. The message can be changed through Moodle language customization.