Deny direct access if the meeting is not available
Up to this patch, the check if a meeting is available or not was ineffective. If the user just replaced /mod/zoom/view.php with /mod/zoom/loadmeeting.php in the URL, he could join the meeting anytime. This patch solves this minor security flaw by doing another availability check on loadmeeting.php.
If the meeting is not available, the user is redirected to the activity overview page again.