Fix debug messages and flaws which were shown if a meeting does not exist
This patch fixes debug messages and flaws which were introduced in commit c22c108 / issue #207 and which were shown if a meeting did not exist on Zoom anymore but the activity overview page showed the unavailability note anyway.

With this patch, the unavailability note is only shown if the meeting exists on Zoom, i.e. the recreate message is not shown.
It also changes the implementation of the recreate message on-the-fly to use the constants from the notification class.