Check if a Zoom account is active before offering it as alternative host
This patch fixes a flaw which existed before when the list of alternative hosts was composed for the alternative hosts user picker.
Adding a pending or inactive user as alternative host will result in a Zoom API error, that's why we check if the user is active before offering it as alternative host.